backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-jose[cryptography]
bcrypt
pydantic-settings
python-dotenv
